Initialize ERC20 token contract as part of starting ParityInstance
This makes it symmetric with BitcoindInstance.
