--- api_tests/lib/ledgers/ethereum.ts
@@ -1,4 +1,0 @@
-export interface EthereumNodeConfig {
-    rpc_url: string;
-    tokenContract: string;
-}
